refactor(generate-mindmap): separate read type utils and create prompt utils
